Rename duke to cat
