+++ src/main/java/cat/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package cat;
 
 import java.io.File;
 import java.io.FileInputStream;
