+++ src/main/java/cat/command/ExitCommand.java
@@ -1,8 +1,8 @@
-package duke.command;
+package cat.command;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.ui.Ui;
+import cat.Storage;
+import cat.TaskList;
+import cat.ui.Ui;
 
 /**
  * A command to exit the program.
