+++ src/main/java/cat/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package duke.ui;
+package cat.ui;
 
 import java.io.IOException;
