+++ src/test/java/cat/ParserTest.java
@@ -1,4 +1,4 @@
-package duke;
+package cat;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
